requirements.txt
pandas
PyJWT
flask
flask-cors
sentence-transformers
torch
pytest
